+++ tests/testthat/test-GDALVector-class.R
@@ -639,7 +639,7 @@ test_that("feature write methods work", {
     feat5$date_fld <- NA
     feat5$datetime_fld <- NA
     feat5$time_fld <- NA
-    feat5$binary_fld <- NA
+    feat5$binary_fld <- raw()
     feat5[[geom_fld]] <- "POINT (5 5)"
 
     expect_true(lyr$createFeature(feat5))
@@ -653,8 +653,8 @@ test_that("feature write methods work", {
     feat6$int64_fld <- bit64::NA_integer64_
     feat6$real_fld <- NA_real_
     feat6$str_fld <- NA_character_
-    feat6$date_fld <- NA_real_
-    feat6$datetime_fld <- NA_real_
+    feat6$date_fld <- as.Date(NA_real_)
+    feat6$datetime_fld <- as.POSIXct(NA_real_)
     feat6$time_fld <- NA_character_
     feat6$binary_fld <- raw(0)
     feat6[[geom_fld]] <- "POINT (6 6)"
@@ -815,13 +815,13 @@ test_that("feature write methods work", {
     # not character or raw vector for geom field
     orig <- feat[[geom_fld]]
     feat[[geom_fld]] <- integer(10)
-    expect_error(lyr$setFeature(feat))
+    expect_false(lyr$setFeature(feat))
     feat[[geom_fld]] <- orig
 
     # not character or raw vector for geom field, in list
     orig <- feat[[geom_fld]]
     feat[[geom_fld]] <- list(numeric(10))
-    expect_error(lyr$setFeature(feat))
+    expect_false(lyr$setFeature(feat))
     feat[[geom_fld]] <- orig
 
     # multi-row data frame
@@ -969,6 +969,61 @@ test_that("feature write methods work", {
     deleteDataset(dsn5)
     rm(lyr)
     rm(dsn5)
+})
+
+test_that("feature batch writing works", {
+    f <- system.file("extdata/ynp_fires_1984_2022.gpkg", package="gdalraster")
+    dsn <- file.path(tempdir(), basename(f))
+    file.copy(f, dsn, overwrite = TRUE)
+
+    sql <- "SELECT incid_name, burn_bnd_ac, ig_date, geom
+            FROM mtbs_perims WHERE ig_year > 2010"
+    lyr <- new(GDALVector, dsn, sql)
+
+    # define new layer by modifying the source definition
+    defn <- lyr$getLayerDefn()
+    # define new attribute field
+    defn$burn_bnd_ha <- ogr_def_field("OFTInteger64")
+    # redefine geom field
+    defn$geom <- ogr_def_geom_field("POINT", srs = defn$geom$srs)
+
+    dst_dsn <- tempfile(fileext = ".gpkg")
+    new_lyr <- ogr_ds_create("GPKG", dst_dsn, "mtbs_centroids",
+                             layer_defn = defn, overwrite = TRUE,
+                             return_obj = TRUE)
+
+    d <- lyr$fetch(-1)
+    # create a new data frame of point features
+    d_new <- data.frame(d[, c("FID", "incid_name", "burn_bnd_ac", "ig_date")])
+    # add new calculated attribute field
+    d_new$burn_bnd_ha <- d_new$burn_bnd_ac / 2.471
+    # add a geom field with the centroids
+    perim_centroids <- g_centroid(d$geom)
+    d_new$geom <- g_create("POINT", perim_centroids)
+
+    # batch write
+    expect_no_error(ret <- new_lyr$batchCreateFeature(d_new))
+    expect_vector(ret, logical(), size = 15)
+    expect_true(all(ret))
+
+    # read back
+    new_lyr$open(read_only = TRUE)
+    expect_equal(new_lyr$getName(), "mtbs_centroids")
+    d_new_out <- new_lyr$fetch(-1)
+    expect_equal(nrow(d_new_out), 15)
+    expect_equal(d_new_out$incid_name, d$incid_name)
+    expect_equal(d_new_out$ig_date, d$ig_date)
+    expect_equal(sum(d_new_out$burn_bnd_ha - (d$burn_bnd_ac / 2.471)), 0,
+                 tolerance = 0.01)
+
+    pt_coords <- g_coords(d_new_out$geom)
+    expect_equal(cbind(pt_coords$x, pt_coords$y), perim_centroids,
+                 ignore_attr = TRUE)
+
+    lyr$close()
+    unlink(dsn)
+    new_lyr$close()
+    unlink(dst_dsn)
 })
 
 test_that("get/set metadata works", {